Add factor out confirmation modal
