+++ packages/common/src/ui/components/modals/ConfirmationModal/ConfirmationModal.stories.tsx
@@ -4,7 +4,7 @@ import { Story } from '@storybook/react';
 import { ConfirmationModal } from './ConfirmationModal';
 import { BaseButton } from '../../buttons';
 import { useToggle } from '../../../../hooks';
-import { useConfirmationModal } from './ConfirmationModalProvider';
+import { useConfirmationModal } from './useConfirmationModal';
 
 export default {
   title: 'Modals/ConfirmationModal',
