+++ packages/common/src/ui/components/modals/ConfirmationModal/index.ts
@@ -1,2 +1,4 @@
-export { ConfirmationModal } from './ConfirmationModal';
+export * from './ConfirmationModal';
 export * from './ConfirmationModalProvider';
+export * from './ConfirmationModalContext';
+export * from './useConfirmationModal';
